% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{commits}
\alias{commits}
\title{Prices of 50,000 round cut diamonds.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{full_name}{}
  \item{author}{}
  \item{author_type}{}
  \item{commit_email}{}
  \item{datetime}{}
  \item{sha}{}
  \item{commit_message}{}
}
}
\usage{
commits
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
