% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_commits_get.R
\name{gh_commits_get}
\alias{gh_commits_get}
\title{Get commits for a repo}
\usage{
gh_commits_get(full_names, days_back, ...)
}
\arguments{
\item{full_names}{Vector of repo names (format 'org/repo')}

\item{days_back}{How many days back to look for commits.}

\item{...}{Pass down options to \code{gh::gh()}}
}
\description{
Get commits for a repo
}
\details{
\strong{New Columns}

\describe{

 \item{full_name}{org/repo}
 \item{author}{Author name}
 \item{datetime}{Date time of commit}
 \item{sha}{Unique hash - can be joined on repo_sha from \code{gh_repo_files_get()}}
 \item{commit_message}{The commit message}
}
}
