% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_file_get.R
\name{gh_file_get}
\alias{gh_file_get}
\title{Get files in a repo}
\usage{
gh_file_get(
  repo = "GithubMetrics",
  org = "OpenPharma",
  file = "DESCRIPTION",
  ...
)
}
\arguments{
\item{repo}{The repo name.}

\item{org}{The Github organisation name}

\item{file}{The filename (or path with filename if not in root)}

\item{...}{Pass down options to \code{gh::gh()}}
}
\description{
Get files in a repo
}
