% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_filetype.R
\name{gh_filetype}
\alias{gh_filetype}
\title{Take a file extension and recode it.}
\usage{
gh_filetype(x, extract_extension = FALSE)
}
\arguments{
\item{x}{Thing to be recoded}

\item{extract_extension}{Do you first need to extract bit after period?}
}
\description{
Take a file extension and recode it.
}
