% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_issues_comments_get.R
\name{gh_issues_comments_get}
\alias{gh_issues_comments_get}
\title{Get comments on issues in a repo}
\usage{
gh_issues_comments_get(full_names, days_back, ...)
}
\arguments{
\item{full_names}{Vector of repo names (format 'org/repo')}

\item{days_back}{How many days back to look for commits.}

\item{...}{Pass down options to \code{gh::gh()}}
}
\description{
Get comments on issues in a repo
}
