% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_issues_get.R
\name{gh_issues_get}
\alias{gh_issues_get}
\title{Get issues for a repo}
\usage{
gh_issues_get(full_names, days_back, ...)
}
\arguments{
\item{full_names}{Vector of repo names (format 'org/repo')}

\item{days_back}{How many days back to look for commits.}

\item{...}{Pass down options to \code{gh::gh()}}
}
\description{
Get issues for a repo
}
\details{
\strong{New Columns}

\describe{

 \item{title}{Issue title}
 \item{url}{Link to issue}
 \item{state}{State of the issue (e.g. open or closed)}
 \item{created, updated, closed}{Unique hash - can be joined on repo_sha from \code{gh_repo_files_get()}}
 \item{comments}{The commit message}
 \item{body}{The commit message}
 \item{author}{The commit message}
}
}
