% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_issues_labels_get.R
\name{gh_issues_labels_get}
\alias{gh_issues_labels_get}
\title{Get issues for a repo}
\usage{
gh_issues_labels_get(
  full_names,
  days_back = 30,
  state = c("open", "closed", "all"),
  ...
)
}
\arguments{
\item{full_names}{Vector of repo names (format 'org/repo')}

\item{days_back}{How many days back to look for commits.}

\item{state}{Defaults to open. Search issues based on open, closed or all.}

\item{...}{Pass down options to \code{gh::gh()}}
}
\description{
Get issues for a repo
}
\details{
\strong{New Columns}

\describe{

 \item{url}{URL for the issue}
 \item{label_name}{Name of the label}
 \item{label_color}{Hex colour}
 \item{label_description}{Label description}
}
}
