% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_health_metrics.R
\name{gh_metric_commits_authors_ever}
\alias{gh_metric_commits_authors_ever}
\title{Total number of commit authors for a repo}
\usage{
gh_metric_commits_authors_ever(commits)
}
\arguments{
\item{commits}{A dataframe of commits}
}
\value{
\itemize{
\item \code{full_name}: repo full name
\item \code{authors_ever}: authors of the commits
}
}
\description{
Uses the
columns \code{fullname}, \code{author}.
}
