% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_health_metrics.R
\name{gh_metric_commits_authors_prepost_midpoint}
\alias{gh_metric_commits_authors_prepost_midpoint}
\title{Trend in author engagement}
\usage{
gh_metric_commits_authors_prepost_midpoint(commits)
}
\arguments{
\item{commits}{A dataframe of commits}
}
\value{
\itemize{
\item \code{full_name}: repo full name
\item \code{authors_prepost_midpoint}: Using time from first to last commit as the time scale, what is total authors in second half - authors in first half.
}
}
\description{
Uses the
columns \code{fullname}, \code{author}.
}
