% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_health_metrics.R
\name{gh_metric_commits_days_since_commit}
\alias{gh_metric_commits_days_since_commit}
\title{Days since update}
\usage{
gh_metric_commits_days_since_commit(commits)
}
\arguments{
\item{commits}{A dataframe of commits}
}
\value{
\itemize{
\item \code{full_name}: repo full name
\item \code{days_since_update}: Date last commit - today's date
}
}
\description{
Days since last commit on master
}
\details{
Uses columns \code{fullname}, \code{date}.
}
