% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_health_metrics.R
\name{gh_metric_commits_prepost_midpoint}
\alias{gh_metric_commits_prepost_midpoint}
\title{Trend in quantity of commits}
\usage{
gh_metric_commits_prepost_midpoint(commits)
}
\arguments{
\item{commits}{A dataframe of commits}
}
\value{
\itemize{
\item \code{full_name}: repo full name
\item \code{commits}: total commits
\item \code{commits_prepost_midpoint}: Using time from first to last commit as the time scale, what is total commits in first half - commit in second half.
}
}
\description{
Calculates the difference in commits between first and last half of a projects
calendar life
}
\details{
Uses columns \code{fullname}, \code{date}.
}
