% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_health_metrics.R
\name{gh_metric_issues}
\alias{gh_metric_issues}
\title{Calculate metrics for the issues}
\usage{
gh_metric_issues(issues)
}
\arguments{
\item{issues}{A dataframe of issues}
}
\value{
\itemize{
\item \code{full_name}: repo full name
\item \code{open_issues}: total open issues
\item \code{median_age_open_issue}: median time open issues have been open
\item \code{median_inactivity_open_issue}: median time an open issue has been inactive
}
}
\description{
Calculates metrics of interest for open source health based on the issue
metadata using the
columns \code{fullname}, \code{state}, \code{days_open}, \code{days_no_activity}.
}
