% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_repo_files_get.R
\name{gh_repo_files_get}
\alias{gh_repo_files_get}
\title{Get files in a repo}
\usage{
gh_repo_files_get(repo_commits, only_last_commit = TRUE, ...)
}
\arguments{
\item{repo_commits}{Output from \code{gh_commits_get()}}

\item{only_last_commit}{Only pull files in the most recent commit}

\item{...}{Pass down options to \code{gh::gh()}}
}
\description{
Get files in a repo
}
\details{
\strong{New Columns}

\describe{

 \item{repo}{Maximum VisitDate observed in the VISIT table}
 \item{file}{Maximum AdministeredDate observed in the MEDICATIONADMINISTRATION table}
 \item{sha_repo}{Can join back onto commit hash}
 \item{sha_commit}{hash}
 \item{extension}{The filetype}
 \item{lang}{Estimate of language from extension using \code{GithubMetrics:::gh_filetype()}}
}
}
