% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_repo_search.R
\name{gh_repo_search}
\alias{gh_repo_search}
\title{Get files in a repo}
\usage{
gh_repo_search(
  code = "RocheData",
  organisation = "PHC",
  full_name = NULL,
  custom = "in:file",
  ...
)
}
\arguments{
\item{code}{Output from \code{gh_commits_get()}}

\item{organisation}{Shortcut to add a filter onto a specific organisation}

\item{full_name}{Shortcut to add a filter on to a specific repo}

\item{custom}{Add your own query}

\item{...}{Pass down options to \code{gh::gh()}}
}
\description{
Get files in a repo
}
\details{
\strong{New Columns}

\describe{

 \item{full_name}{org/repo}
 \item{name}{repo name}
 \item{file_name}{File name}
 \item{path}{Path within repo to file including filename}
 \item{url}{URL to the file and commit on github}
 \item{score}{i didn't actually look this up... maybe matching score}
 \item{lang}{Language guessed via \code{GithubMetrics:::gh_filetype()}}
}
}
