% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_repos_clean.R
\name{gh_repos_clean}
\alias{gh_repos_clean}
\title{Flatten repo list into a tibble}
\usage{
gh_repos_clean(orgs)
}
\arguments{
\item{orgs}{Returned object from \code{gh_repos_get()}}
}
\description{
Flatten repo list into a tibble
}
\details{
\strong{New Columns}

\describe{
 \item{name}{Name of the repo}
 \item{full_name}{Org and name of repo}
 \item{size}{Raw size of the repo}
 \item{updated_at}{When it was last updated}
 \item{default_branch}{Name of the default branch}
 \item{language}{Github estimate of main language}
 \item{MB}{Size converted to MB}
}
}
