% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_repos_get.R
\name{gh_repos_get}
\alias{gh_repos_get}
\title{Get org access}
\usage{
gh_repos_get(org, ...)
}
\arguments{
\item{org}{Which Github org are you looking for}

\item{...}{Pass down options to \code{gh::gh()}}
}
\value{
Returns gh::gh_response object which contains repos in the org
}
\description{
Get org access
}
