% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_score.R
\name{gh_score}
\alias{gh_score}
\title{Score the open source health of the repos}
\usage{
gh_score(x)
}
\arguments{
\item{x}{Out of joining the metrics functions}
}
\value{
\itemize{
\item \code{full_name}: repo full name
\item \code{warnings}: A list of where this scored badly
\item \code{score}: The OS score
}
}
\description{
Score the open source health of the repos
}
\details{
This function will apply the following scoring metric (positive answer = 1 point, negative is 0 points):
\itemize{
\item has the default branch had an commit in the last 6 months
\item has the project had > 25 commits
\item does it have more than 5 unique contributors
\item the open issues have a median time from creation of <6 months
\item for the open issues, there is a median time since comment of <3 months
\item the project is more active in commits in the last half of it's existance than the first
\item the project has more contributors active in the last half of it's existance then the first
}
}
