% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_user_events_get.R
\name{gh_user_event_get}
\alias{gh_user_event_get}
\title{Get a users activity in last 90 days}
\usage{
gh_user_event_get(user, ...)
}
\arguments{
\item{user}{The github username}

\item{...}{Pass down to \code{gh}}
}
\description{
If Authenticated, and you ask for yourself (the authenticated user) you will
also get private activity. Otherwise it's just public.
}
