% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_user_get.R
\name{gh_user_get}
\alias{gh_user_get}
\title{Get files in a repo}
\usage{
gh_user_get(user, verbose = FALSE, ...)
}
\arguments{
\item{user}{Character vector with Github username (called author in their objs)}

\item{verbose}{Print out updates on what is being pulled?}

\item{...}{Pass down options to \code{gh::gh()}}
}
\description{
Get files in a repo
}
