% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_user_repos_external_get.R
\name{gh_user_repos_external_get}
\alias{gh_user_repos_external_get}
\title{Get repos interacted with outside user org}
\usage{
gh_user_repos_external_get(user, token = Sys.getenv("GITHUB_PAT"))
}
\arguments{
\item{user}{The github username}

\item{token}{Github token}
}
\description{
It will return a vector or repos. It ONLY pulls first 100 currently
as no pagination in place.
}
