% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_weekly_codelines.R
\name{gh_weekly_code_lines_get}
\alias{gh_weekly_code_lines_get}
\alias{gh_weekly_codelines_get}
\title{Get weekly number of lines of code from Github repos}
\usage{
gh_weekly_codelines_get(full_names, year_for_stats, ...)
}
\arguments{
\item{full_names}{A character vector containing names of repositories (format 'org/repo').}

\item{year_for_stats}{A character year to filter statistics.}

\item{...}{Pass down options to \code{gh::gh()}}
}
\value{
A list of repo stats, with every slot (repo) containing: \itemize{
  \item{Starting date of a week.} \item{Lines of code added in a given week.}
  \item{Lines of code removed in a given week.}}
}
\description{
This function is a handy wrapper to iterate requests for all repos
  of a given Github account to the Github endpoint
  \code{https://api.github.com/repos/OWNER/REPO/stats/code_frequency} For
  more information on the endpoint you can visit:\cr\cr
  \link{https://docs.github.com/en/rest/metrics/statistics?apiVersion=2022-11-28#get-the-weekly-commit-activity}
}
