% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_weekly_codelines.R
\name{gh_weekly_codelines_dt}
\alias{gh_weekly_codelines_dt}
\title{Weekly number of added and removed code lines from Github repos}
\usage{
gh_weekly_codelines_dt(
  full_names,
  year_for_stats,
  unit = c("none", "thous", "mln", "bln"),
  ...
)
}
\arguments{
\item{full_names}{A character vector containing names of repositories (format
'org/repo').}

\item{year_for_stats}{A character year to filter statistics.}

\item{unit}{A unit to show values for added and removed lines.}

\item{...}{Pass down options to \code{gh::gh()}}
}
\value{
A tibble with columns of added and removed lines numbers per week.
}
\description{
This function prepares table for weekly number of lines of code
  added and removed per week from all repositories from a Github account.
}
\examples{
\dontrun{
repos_openpharma <- gh_repos_get(org = "openpharma") \%>\%
  gh_repos_clean()

gh_weekly_stats_dt <- gh_weekly_codelines_dt(full_names = repos_openpharma$full_name,
                                             year_of_stats = "2022",
                                             unit = "thous")
}

}
