% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_weekly_commits.R
\name{gh_weekly_commits_dt}
\alias{gh_weekly_commits_dt}
\title{Prepare table for Github weekly number of commits}
\usage{
gh_weekly_commits_dt(fullnames, ...)
}
\arguments{
\item{...}{Pass down options to \code{gh::gh()}}

\item{full_names}{A character vector containing names of repositories (format
'org/repo').}
}
\value{
A dataframe with number of commits per week for last 52 weeks.
}
\description{
Prepare table for Github weekly number of commits
}
\examples{
\dontrun{
 repos_openpharma <- gh_repos_get(org = "openpharma") \%>\%
  gh_repos_clean()

 gh_weekly_stats <- gh_weekly_commits_dt(full_names = repos_openpharma$full_name)
}

}
