% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_weekly_commits.R
\name{gh_weekly_commits_get}
\alias{gh_weekly_commits_get}
\title{Get number of commits per week from a Github account}
\usage{
gh_weekly_commits_get(full_names, ...)
}
\arguments{
\item{full_names}{A character vector containing names of repositories (format 'org/repo').}

\item{...}{Pass down options to \code{gh::gh()}}
}
\value{
A list of repo stats, with every slot (repo) containing number of
  commits per week.
}
\description{
This function captures number of commits from all repos of a
  given Github account from last 52 weeks
}
\details{
This function is a handy wrapper to iterate requests for all repos
  of a given Github account to the Github endpoint
  \code{https://api.github.com/repos/OWNER/REPO/stats/participation} For more
  information on the endpoint you can visit:\cr\cr
  \link{https://docs.github.com/en/rest/metrics/statistics?apiVersion=2022-11-28#get-the-weekly-commit-count}
}
