% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GithubMetrics-package.R
\docType{package}
\name{GithubMetrics-package}
\alias{GithubMetrics}
\alias{GithubMetrics-package}
\title{GithubMetrics: Quickly get key metrics on Github repositaries}
\description{
The aim of this package is to provide a wrapper on gh to quickly get you key Github repo metadata you need.
}
\author{
\strong{Maintainer}: James Black \email{james@epijim.uk}

}
\keyword{internal}
