% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{issues}
\alias{issues}
\title{Prices of 50,000 round cut diamonds.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{full_name}{}
  \item{issue_number}{}
  \item{state}{}
  \item{title}{}
  \item{url}{}
  \item{created}{}
  \item{updated}{}
  \item{closed}{}
  \item{comments}{}
  \item{body}{}
  \item{author}{}
  \item{author_status}{}
}
}
\usage{
issues
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
