% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GithubMetrics-package.R
\docType{package}
\name{GithubMetrics-package}
\alias{GithubMetrics}
\alias{GithubMetrics-package}
\title{GithubMetrics: Quickly get key metrics on Github repositaries}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: James Black \email{james.black.jb2@roche.com}

}
\keyword{internal}
